+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+## 0.14.10
+
 ## 0.14.9
 ### Features
 - Custom error page.